requirements.txt
fastapi
uvicorn
pydantic
requests
python-multipart

langchain==0.1.14
langchain-community==0.0.32
langchain-core==0.1.53
langchain-huggingface==0.0.3
langchain-google-genai==0.0.5
langchain-text-splitters==0.0.2
langsmith==0.1.147

sentence-transformers
transformers
faiss-cpu
pdfplumber
python-docx
python-dotenv

nest_asyncio  
router.py
fastapi 
uvicorn  
protobuf==3.20.3